requirements.txt
# connectors
elasticsearch>=7.13.1
mysql-connector-python>=8.0.19
pandas>=1.0.1
psycopg2>=2.8.6
pymongo>=3.10.1
python-dotenv>=0.17.1
requests>=2.25.1
sqlalchemy>=1.3.16
tqdm>=4.43.0

# mysql
mysql-connector-python>=8.0.19
numpy>=1.20.3
pandas>=1.0.1
python-dateutil>=2.8.1
python-dotenv>=0.17.1
requests>=2.25.1
text-unidecode>=1.3
tqdm>=4.43.0

# mongo
pymongo>=3.10.1
python-dotenv>=0.17.1
requests>=2.25.1

# elastic
elasticsearch>=7.13.1
python-dotenv>=0.17.1
requests>=2.25.1
tqdm>=4.43.0

# postgres
psycopg2>=2.8.6
python-dotenv>=0.17.1
requests>=2.25.1

# api
babel>=2.9.0
beautifulsoup4>=4.9.1
dnspython>=2.0.0
elasticsearch>=7.13.1
lxml>=4.5.2
pendulum>=2.0.0
phonenumbers>=8.11.5
psutil>=5.7.2
pycountry>=19.8.18
pymailcheck>=1.0.0
pymongo>=3.10.1
python-dateutil>=2.8.1
python-dotenv>=0.17.1
requests>=2.25.1
text-unidecode>=1.3
tqdm>=4.43.0

# address
psutil>=5.7.2
pycountry>=19.8.18
python-dotenv>=0.17.1
requests>=2.25.1

# email
babel>=2.9.0
beautifulsoup4>=4.9.1
dnspython>=2.0.0
elasticsearch>=7.13.1
pendulum>=2.0.0
phonenumbers>=8.11.5
psutil>=5.7.2
pycountry>=19.8.18
pymailcheck>=1.0.0
pymongo>=3.10.1
python-dateutil>=2.8.1
python-dotenv>=0.17.1
requests>=2.25.1
text-unidecode>=1.3
tqdm>=4.43.0

# phone
beautifulsoup4>=4.9.1
elasticsearch>=7.13.1
lxml>=4.5.2
pendulum>=2.0.0
phonenumbers>=8.11.5
psutil>=5.7.2
pycountry>=19.8.18
pymongo>=3.10.1
python-dotenv>=0.17.1
requests>=2.25.1
tqdm>=4.43.0

# ddt
matplotlib>=3.2.0
mysql-connector-python>=8.0.19
numpy>=1.20.3
openpyxl>=3.0.3
pandas>=1.0.1
python-docx>=0.8.10
python-dotenv>=0.17.1
requests>=2.25.1
seaborn>=0.10.1
xlrd>=1.2.0
xlsxwriter>=1.2.8

# env
python-dotenv>=0.17.1

# handlers
tqdm>=4.43.0

# parsers
numpy>=1.20.3
pandas>=1.0.1
python-dateutil>=2.8.1
text-unidecode>=1.3

# persondata
beautifulsoup4>=4.9.1
dnspython>=2.0.0
elasticsearch>=7.13.1
lxml>=4.5.2
numpy>=1.20.3
pandas>=1.0.1
pendulum>=2.0.0
phonenumbers>=8.11.5
psutil>=5.7.2
pycountry>=19.8.18
pymailcheck>=1.0.0
pymongo>=3.10.1
python-dotenv>=0.17.1
python-dateutil>=2.8.1
requests>=2.25.1
text-unidecode>=1.3
tqdm>=4.43.0

# persons
babel>=2.9.0
beautifulsoup4>=4.9.1
dnspython>=2.0.0
elasticsearch>=7.13.1
pendulum>=2.0.0
phonenumbers>=8.11.5
psutil>=5.7.2
pycountry>=19.8.18
pymailcheck>=1.0.0
pymongo>=3.10.1
python-dotenv>=0.17.1
python-dateutil>=2.8.1
requests>=2.25.1
text-unidecode>=1.3
tqdm>=4.43.0

# platform
paramiko>=2.7.2
pycryptodome<4.0.0
pymongo>=3.10.1
python-dotenv>=0.17.1
requests>=2.25.1

# requests
psutil>=5.7.2
python-dotenv>=0.17.1
requests>=2.25.1

# secrets
python-dotenv>=0.17.1
requests>=2.25.1

# visualizations
matplotlib>=3.2.0
numpy>=1.20.3
seaborn>=0.11.1

# tests
hypothesis>=5.6.0
pytest>=5.3.5

# geo
Shapely>=1.7.1
